fix(build): local deploy script supports config

+++ deploy_fresh_test_site.sh
@@ -3,6 +3,13 @@
 DEFAULT_SUPERUSER_PASSWORD="My sup3r p4ssw0rd!"
 CURRENT_USER=`whoami`
 
+die () {
+	echo >&2 "$@"
+	exit 1
+}
+[ "$#" -eq 1 ] || die "You must provide the path to the JSON secrets file"
+[[ -f "$1" ]] || die "Can't find that file: $1"
+
 # Recreate virtual environment and install site dependencies
 rm -rfv ./venv
 (set -x; python3 -m venv venv)
@@ -10,6 +17,7 @@ rm -rfv ./venv
 ./venv/bin/pip install --upgrade tomli
 ./venv/bin/pip install --upgrade `./venv/bin/python -c "import tomli; tf = open('pyproject.toml', 'rb'); c = tomli.load(tf); print(' '.join(c['project']['dependencies']))"`
 ./venv/bin/pip install --upgrade `./venv/bin/python -c "import tomli; tf = open('pyproject.toml', 'rb'); c = tomli.load(tf); print(' '.join(c['build-system']['requires']))"`
+./venv/bin/pip install --upgrade `./venv/bin/python -c "import tomli; tf = open('pyproject.toml', 'rb'); c = tomli.load(tf); print(' '.join(c['project']['optional-dependencies']['dev']))"`
 
 # Create site structure
 rm -rfv ./test_site
@@ -20,7 +28,7 @@ rm -rfv ./test_site
 (set -x; ln -s ../../settings.py ./test_site/test_site/local_settings.py)
 
 # DB handling
-(set -x; ./venv/bin/python ./test_site/manage.py migrate --settings=test_site.local_settings)
+(set -x; env `./venv/bin/python -m env_pipes vars_from_file --uppercase_vars $1` ./venv/bin/python ./test_site/manage.py migrate --settings=test_site.local_settings)
 
 # Create super user
 (export DJANGO_SUPERUSER_LOGIN=$CURRENT_USER
@@ -28,11 +36,11 @@ export DJANGO_SUPERUSER_FIRSTNAME="$DJANGO_SUPERUSER_LOGIN"
 export DJANGO_SUPERUSER_LASTNAME="$DJANGO_SUPERUSER_LOGIN"
 export DJANGO_SUPERUSER_EMAIL="$DJANGO_SUPERUSER_LOGIN@invalid.local"
 export DJANGO_SUPERUSER_PASSWORD=$DEFAULT_SUPERUSER_PASSWORD
-./venv/bin/python ./test_site/manage.py createsuperuser --noinput --settings=test_site.local_settings)
+env `./venv/bin/python -m env_pipes vars_from_file --uppercase_vars $1` ./venv/bin/python ./test_site/manage.py createsuperuser --noinput --settings=test_site.local_settings)
 
 echo "Now run with:
 
-env DJANGO_DEBUG=true ./venv/bin/python ./test_site/manage.py runserver --settings=test_site.local_settings
+env DJANGO_DEBUG=true \`./venv/bin/python -m env_pipes vars_from_file --uppercase_vars $1\` ./venv/bin/python ./test_site/manage.py runserver --settings=test_site.local_settings
 
 Then go to http://localhost:8000/admin and use credentials $CURRENT_USER:$DEFAULT_SUPERUSER_PASSWORD
 "
